more configuration changes
