+++ tsconfig.json
@@ -12,7 +12,11 @@
     "preserveConstEnums": true,
     "sourceMap": true,
     "skipLibCheck": true,
+    "jsx": "react",
     "baseUrl": ".",
+    "paths": {
+      "!/*": ["./*"]
+    },
     "typeRoots": ["./node_modules/@types"],
     "lib": ["dom", "es2015", "es2016"],
     "outDir": "lib/"
